Day 18: Make it work correctly on python3 (not just pypy3)
Here is a notable difference between pypy3 (5.9.0-beta0/Python 3.5.3) and
python3 (Python 3.6.3): When a generator is no longer used (e.g. when it
runs out of scope), python3 will call .close() (or otherwise raise
GeneratorExit within the generators), whereas pypy3 will not. Thus, we
cannot use on GeneratorExit to trigger the printing of sent values, since
we end up printing too much on python3.

Instead, introduce an explicit PrintSent exception that can be raised
within a 'thread' to trigger it to print its sent value.